extract `RegistrationForm` component - convert css styles to sass modules
