+++ src/components/forms/RegisterForm.module.scss
@@ -1,0 +1,18 @@
+.register {
+    &__form {
+        padding: 32px 16px 8px 16px;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        gap: 24px;
+    }
+
+    &__buttons {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        gap: 6px;
+    }
+}
