+++ src/components/forms/index.ts
@@ -1,5 +1,7 @@
 import LoginForm from './LoginForm';
+import RegisterForm from './RegisterForm';
 
 export {
-  LoginForm
+  LoginForm,
+  RegisterForm
 };
