--- src/pages/register/RegisterPage.css
@@ -1,48 +1,0 @@
-.register__body {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100vh;
-  background: radial-gradient(#9bf268, #389602);
-}
-
-.register__form {
-  padding: 32px 16px 8px 16px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 24px;
-}
-
-.register__form input {
-  font-size: 24px;
-  border-style: solid;
-  border-color: #bcbcbc;
-  border-width: 1px;
-  border-radius: 10px;
-  padding: 8px 16px;
-  width: 100%;
-  -moz-box-shadow: inset 0 3px 6px rgba(0, 0, 0, 0.16);
-  -webkit-box-shadow: inset 0 3px 6px rgba(0, 0, 0, 0.16);
-  box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.16);
-}
-
-.register__form input::placeholder {
-  color: #bcbcbc;
-  opacity: 1;
-}
-
-.register__buttons {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 6px;
-}
-
-.register__buttons a {
-  color: #707070;
-}
