+++ src/pages/register/styles.module.scss
@@ -1,0 +1,21 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100vh;
+    background: radial-gradient(#9bf268, #389602);
+}
+
+.register {
+    background-color: white;
+    display: flex; 
+    flex-direction: column;
+    align-items: center;
+    gap: 16px;
+
+    &__link {
+        color: #707070;
+    }
+}
